requirements.txt
pandas>=0.25.3
numpy
matplotlib
python-dotenv>=0.14.0  # Manage enviroment vars
